feat(routes): move or routes into org module
- update put method to patch
- update routes package
